[Issue-2710] Improve Mission Pool extension
