+++ packages/extension-koni-ui/src/utils/index.ts
@@ -229,3 +229,4 @@ export * from './crowdloan';
 export * from './crowdloan';
 export * from './history';
 export * from './earning';
+export * from './missionPools';
